requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
SQLAlchemy==2.0.36
pydantic-settings==2.6.1
PyJWT==2.10.1
passlib[argon2]==1.7.4
argon2-cffi==23.1.0
httpx==0.27.2
email-validator==2.2.0
requests